+++ mail_2_rmq.py
@@ -68,24 +68,33 @@
 """
 
 # Libraries and Global Variables
+from __future__ import print_function
+from __future__ import absolute_import
 
 # Standard
 from __future__ import print_function
 import sys
 import os
 import datetime
-
-# Third-party
 import email.Parser
 import re
 import base64
+import io
 
 # Local
-import lib.arg_parser as arg_parser
-import lib.gen_libs as gen_libs
-import lib.gen_class as gen_class
-import rabbit_lib.rabbitmq_class as rabbitmq_class
-import version
+try:
+    from .lib import arg_parser
+    from .lib import gen_libs
+    from .lib import gen_class
+    from .rabbit_lib import rabbitmq_class
+    from . import version
+
+except (ValueError, ImportError) as err:
+    import lib.arg_parser as arg_parser
+    import lib.gen_libs as gen_libs
+    import lib.gen_class as gen_class
+    import rabbit_lib.rabbitmq_class as rabbitmq_class
+    import version
 
 __version__ = version.__version__
 
@@ -310,10 +319,10 @@ def process_attach(msg, log, cfg):
             if item.get_content_type() in cfg.attach_types:
                 tname = os.path.join(cfg.tmp_dir, item.get_filename())
                 log.log_info("Attachment detected: %s" % (item.get_filename()))
-                open(tname, "wb").write(item.get_payload(decode=True))
+                io.open(tname, "wb").write(item.get_payload(decode=True))
                 fname = tname + ".encoded"
                 fname_list.append(fname)
-                base64.encode(open(tname, 'rb'), open(fname, 'wb'))
+                base64.encode(io.open(tname, 'rb'), io.open(fname, 'wb'))
                 err_flag, err_msg = gen_libs.rm_file(tname)
 
                 if err_flag:
@@ -476,7 +485,7 @@ def process_message(cfg, log):
     if subj in cfg.valid_queues:
         process_subj(cfg, log, subj, msg)
 
-    elif from_addr and from_addr in cfg.queue_dict.keys():
+    elif from_addr and from_addr in list(cfg.queue_dict.keys()):
         process_from(cfg, log, msg, from_addr)
 
     else: