+++ test/unit/mail_2_rmq/convert_bytes.py
@@ -1,11 +1,11 @@
 # Classification (U)
 
-"""Program:  check_nonprocess.py
+"""Program:  convert_bytes.py
 
-    Description:  Unit testing of check_nonprocess in mail_2_rmq.py.
+    Description:  Unit testing of convert_bytes in mail_2_rmq.py.
 
     Usage:
-        test/unit/mail_2_rmq/check_nonprocess.py
+        test/unit/mail_2_rmq/convert_bytes.py
 
     Arguments:
 
@@ -21,7 +21,6 @@ import unittest
 # Local
 sys.path.append(os.getcwd())
 import mail_2_rmq
-import lib.gen_libs as gen_libs
 import version
 
 __version__ = version.__version__
@@ -35,7 +34,7 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp
-        test_check_nonprocess
+        test_conversion
 
     """
 
@@ -49,20 +48,23 @@ class UnitTest(unittest.TestCase):
 
         """
 
-        pass
+        self.data = "Data String"
+        self.results2 = "Data String"
+        self.results3 = b"Data String"
 
-    def test_check_nonprocess(self):
+    def test_conversion(self):
 
-        """Function:  test_check_nonprocess
+        """Function:  test_conversion
 
-        Description:  Test processing of non-processed emails.
+        Description:  Test conversion based on Python version.
 
         Arguments:
 
         """
 
-        with gen_libs.no_std_out():
-            self.assertFalse(mail_2_rmq.check_nonprocess("", ""))
+        results = self.results2 if sys.version_info < (3, 0) else self.results3
+
+        self.assertEqual(mail_2_rmq.convert_bytes(self.data), results)
 
 
 if __name__ == "__main__":