+++ test/unit/sonarqube_code_coverage.sh
@@ -11,6 +11,7 @@ echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/archive_email.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/check_nonprocess.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/connect_process.py
+coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/convert_bytes.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/filter_subject.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/get_text.py
 coverage run -a --source=mail_2_rmq test/unit/mail_2_rmq/help_message.py