+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ### Changed
 - Converted imports to use Python 2.7 or Python 3.
+- process_message: Changed output of dictionary keys call to a list.
+- process_attach: Replaced open() call with io.open() call.
 
 
 ## [1.3.1] - 2021-11-19