+++ CHANGELOG.md
@@ -4,6 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [1.3.2] - 2022-12-09
+- Updated to work in Python 3 too
+- Upgraded python-lib to v2.9.4
+- Upgraded rabbitmq-lib to v2.2.1
+
+### Changed
+- Converted imports to use Python 2.7 or Python 3.
+- process_message: Changed output of dictionary keys call to a list.
+- process_attach: Replaced open() call with io.open() call and called convert_bytes to convert data to bytes if in Python 3.
+- parse_email: Using different email parsers depending on Python version.
+
+### Added
+- convert_bytes: Converts a string to bytes if in a Python 3 environment.
+
+
 ## [1.3.1] - 2021-11-19
 - Upgrade python-lib to v2.9.2
 - Upgrade rabbitmq-lib to v2.2.0